Setup App component
Data originates from the main component.

Require the Orders module. It's not actually a React component at all,
so this ought to be moved elsewhere in the future...

Remove the 'maxRange' property from the settings object. Add a general
method, 'genPoints', which is a general method you can use with an Order
callback. Add a 'setPoints' method, which sets the point property to an
object, mapping an array to the name of a corresponding callback function
from the Orders module.